requirements.txt.txt
aiogram>=3.15.0
apscheduler
aiohttp
aiosqlite
python-dotenv